audio: multiband_drc: move out multiband drc user header file
Since there is only one file for user folder, move it out and delete
this folder directly.
